// ScheduleMeFrontEnd/src/app/shared/openapi-api-client/model/authenticateResponse.ts
/**
 * scheduleme.server | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthenticateResponse { 
    id?: string | null;
    title?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    phoneNumber?: string | null;
    dob: string | null;
    role?: string | null;
    created?: string;
    updated?: string | null;
    isVerified?: boolean;
    jwtToken?: string | null;
}